+++ src/sys/class/alo/session/sqlsession.php
@@ -20,13 +20,6 @@
    class SQLSession extends AbstractSession {
 
       /**
-       * The session data table
-       *
-       * @var string
-       */
-      const TABLE_NAME = 'alo_session';
-
-      /**
        * Reference to database instance
        *
        * @var MySQL
@@ -50,7 +43,7 @@
 
       protected function write() {
          $this->db->prepQuery('REPLACE INTO `'
-            . self::TABLE_NAME . '`(`id`,`data`,`access`) VALUES('
+            . ALO_SESSION_TABLE_NAME . '`(`id`,`data`,`access`) VALUES('
             . '?,?,CURRENT_TIMESTAMP)', [
                $this->id,
                json_encode($this->data)
@@ -62,7 +55,7 @@
 
       protected function fetch() {
          $sql = $this->db->prepQuery('SELECT `data` '
-            . 'FROM `' . self::TABLE_NAME . '` '
+            . 'FROM `' . ALO_SESSION_TABLE_NAME . '` '
             . 'WHERE `id`=? '
             . 'LIMIT 1', [$this->id], [
                MySQL::V_CACHE => false
@@ -78,7 +71,7 @@
       }
 
       function terminate() {
-         $this->db->prepQuery('DELETE FROM `' . self::TABLE_NAME . '` '
+         $this->db->prepQuery('DELETE FROM `' . ALO_SESSION_TABLE_NAME . '` '
             . 'WHERE `id`=? '
             . 'LIMIT 1', [$this->id], [MySQL::V_CACHE => false]);