+++ tests/SQLSessionTest.php
@@ -1,0 +1,31 @@
+<?php
+
+   namespace Alo\Session;
+
+   class SQLSessionTest extends \PHPUnit_Framework_TestCase {
+
+      /**
+       * @dataProvider definedProvider
+       */
+      function testDefined($key) {
+         $this->assertTrue(defined($key), $key . ' wasn\'t defined');
+      }
+
+      function definedProvider() {
+         return [
+            ['ALO_SESSION_CLEANUP'],
+            ['ALO_SESSION_TIMEOUT'],
+            ['ALO_SESSION_COOKIE'],
+            ['ALO_SESSION_FINGERPRINT'],
+            ['ALO_SESSION_MC_PREFIX'],
+            ['ALO_SESSION_TABLE_NAME']
+         ];
+      }
+
+      static function sess() {
+         if (!\Alo::$session) {
+            \Alo::$session = new SQLSession();
+         }
+      }
+   }
+